+++ server/app.ts
@@ -5,7 +5,6 @@ import { cors } from 'hono/cors'
 import type { Auth } from './auth'
 import { createDeps } from './composition'
 import backgroundJobs from './http/background-jobs'
-import { cloudStore, cloudStoreWebhooks } from './http/cloud-store'
 import downloadTasks from './http/downloads/download-tasks'
 import downloaders, { downloaderSelfRoute } from './http/downloads/downloaders'
 import { events } from './http/events'
@@ -27,6 +26,7 @@ import { adminInviteCodes, publicInviteCodes } from './http/site/invite-codes'
 import { licensing, licensingAdmin } from './http/site/licensing'
 import storages from './http/site/storages'
 import system from './http/site/system'
+import { cloudStore, cloudStoreWebhooks } from './http/store'
 import { adminTeams, publicTeams, teams } from './http/teams'
 import trash from './http/trash'
 import { users } from './http/users'
