--- server/http/cloud-store.ts
@@ -1,2 +1,0 @@
-export { cloudStore } from './store/storefront'
-export { cloudStoreWebhooks } from './store/webhooks'
