+++ server/http/store/webhooks.ts
@@ -1,7 +1,7 @@
 import { Hono } from 'hono'
 import type { Env } from '../../middleware/platform'
 import { requireFeature } from '../../middleware/require-feature'
-import { processDeliveryWebhook } from '../../usecases/store/cloud-store'
+import { processDeliveryWebhook } from '../../usecases/store/store'
 import { getCloudBaseUrl, parseJson, sha256Hex } from './helpers'
 
 export const cloudStoreWebhooks = new Hono<Env>().use(requireFeature('quota_store')).post('/webhook', async (c) => {
